+++ cvpods/configs/base_classification_config.py
@@ -34,9 +34,8 @@ _config_dict = dict(
             DEEP_STEM=False,
         ),
     ),
-    INPUT=dict(
-        FORMAT="RGB",
-    ),
+    INPUT=dict(FORMAT="RGB"),
+    TRAINER=dict(TYPE="EpochRunner"),
     SOLVER=dict(
         IMS_PER_DEVICE=32,  # defalut: 8 gpus x 32 = 256
     ),
