+++ cvpods/engine/__init__.py
@@ -4,6 +4,6 @@ from .hooks import *
 from .launch import *
 from .predictor import *
 from .setup import *
-from .trainer import *
+from .runner import *
 
 __all__ = [k for k in globals().keys() if not k.startswith("_")]
