+++ cvpods/utils/dump/events.py
@@ -116,7 +116,7 @@ class TensorboardXWriter(EventWriter):
     Write all scalars to a tensorboard file.
     """
 
-    def __init__(self, log_dir: str, window_size: int = 20, **kwargs):
+    def __init__(self, log_dir: str, window_size: int = 1, **kwargs):
         """
         Args:
             log_dir (str): the directory to save the output events
@@ -151,7 +151,7 @@ class CommonMetricPrinter(EventWriter):
     To print something different, please implement a similar printer by yourself.
     """
 
-    def __init__(self, max_iter):
+    def __init__(self, max_iter, window_size=20, **kwargs):
         """
         Args:
             max_iter (int): the maximum number of iterations to train.
@@ -159,14 +159,23 @@ class CommonMetricPrinter(EventWriter):
         """
         self.logger = logging.getLogger(__name__)
         self._max_iter = max_iter
+        if "epoch" in kwargs:
+            self._epoch = kwargs["epoch"]
+            self._epoch_iters = self._max_iter // self._epoch if self._epoch is not None else None
+        else:
+            self._epoch = None
+            self._epoch_iters = None
+
+        self._window_size = window_size
         self._last_write = None
 
-    def write(self):
+    def write(self, window_size=20):
+
         storage = get_event_storage()
         iteration = storage.iter
 
         try:
-            data_time = storage.history("data_time").avg(20)
+            data_time = storage.history("data_time").avg(self._window_size)
         except KeyError:
             # they may not exist in the first few iterations (due to warmup)
             # or when SimpleTrainer is not used
@@ -203,30 +212,35 @@ class CommonMetricPrinter(EventWriter):
         # NOTE: max_mem is parsed by grep in "dev/parse_results.sh"
         losses = "  ".join(
             [
-                "{}: {:.3f}".format(k, v.median(20))
+                "{}: {:.3f}".format(k, v.median(self._window_size))
                 for k, v in storage.histories().items()
                 if "loss" in k
             ]
         )
         other_metrics = "  ".join(
             [
-                "{}: {:.3f}".format(k, v.median(20))
+                "{}: {:.3f}".format(k, v.median(self._window_size))
                 for k, v in storage.histories().items()
                 if "loss" not in k and k not in ["data_time", "time", "lr"] and "/" not in k
             ]
         )
+
         self.logger.info(
-            ("eta: {eta}  iter: {iter}/{max_iter}  {losses}  {other_metrics}  {time}  "
+            ("eta: {eta}  {name}: {iter}|{max_iter}  {losses}  {other_metrics}  {time}  "
              "{data_time}  lr: {lr}  {memory}").format(
-                eta=eta_string,
-                iter=iteration + 1,
-                max_iter=self._max_iter,
-                losses=losses,
-                other_metrics=other_metrics,
-                time="time: {:.4f}".format(iter_time) if iter_time is not None else "",
-                data_time="data_time: {:.4f}".format(data_time) if data_time is not None else "",
-                lr=lr,
-                memory="max_mem: {:.0f}M".format(max_mem_mb) if max_mem_mb is not None else "",
+                 eta=eta_string,
+                 name='epoch|iter' if self._epoch_iters is not None else 'iter',
+                 iter=iteration + 1 if self._epoch_iters is None
+                 else f"[{(iteration + 1) // self._epoch_iters + 1}"
+                 f"/{self._max_iter // self._epoch_iters}]",
+                 max_iter=self._max_iter if self._epoch_iters is None
+                 else f"[{(iteration + 1) % self._epoch_iters}/{self._epoch_iters}]",
+                 losses=losses,
+                 other_metrics=other_metrics,
+                 time="time: {:.4f}".format(iter_time) if iter_time is not None else "",
+                 data_time="data_time: {:.4f}".format(data_time) if data_time is not None else "",
+                 lr=lr,
+                 memory="max_mem: {:.0f}M".format(max_mem_mb) if max_mem_mb is not None else "",
             )
         )
 
@@ -238,7 +252,7 @@ class EventStorage:
     In the future we may add support for storing / logging other types of data if needed.
     """
 
-    def __init__(self, start_iter=0):
+    def __init__(self, start_iter=0, window_size=20):
         """
         Args:
             start_iter (int): the iteration number to start with
@@ -246,6 +260,7 @@ class EventStorage:
         self._history = defaultdict(HistoryBuffer)
         self._smoothing_hints = {}
         self._latest_scalars = {}
+        self._window_size = window_size
         self._iter = start_iter
         self._current_prefix = ""
         self._vis_data = []
@@ -345,7 +360,7 @@ class EventStorage:
         result = {}
         for k, v in self._latest_scalars.items():
             result[k] = (
-                self._history[k].median(window_size) if self._smoothing_hints[k] else v
+                self._history[k].median(self._window_size) if self._smoothing_hints[k] else v
             )
         return result
