+++ tools/train_net.py
@@ -17,35 +17,25 @@ You may want to write your own script with your datasets and other customization
 """
 import logging
 import os
-import pickle as pkl
-import sys
 from collections import OrderedDict
 from colorama import Fore, Style
 
-import torch
-
-from cvpods.checkpoint import DefaultCheckpointer
-from cvpods.engine import DefaultTrainer, default_argument_parser, default_setup, hooks, launch
-from cvpods.evaluation import build_evaluator, verify_results
+from cvpods.engine import RUNNERS, default_argument_parser, default_setup, hooks, launch
 from cvpods.modeling import GeneralizedRCNNWithTTA
-from cvpods.utils import comm
 
 from config import config  # noqa: E402
 from net import build_model  # noqa: E402
 
-sys.path.insert(0, '.')
 
-
-class Trainer(DefaultTrainer):
+def trainer_decrator(cls):
     """
-    We use the "DefaultTrainer" which contains pre-defined default logic for
+    We use the "DefaultRunner" which contains pre-defined default logic for
     standard training workflow. They may not work for you, especially if you
     are working on a new research project. In that case you can use the cleaner
-    "SimpleTrainer", or write your own training loop. You can use
+    "SimpleRunner", or write your own training loop. You can use
     "tools/plain_train_net.py" as an example.
     """
 
-    @classmethod
     def build_evaluator(cls, cfg, dataset_name, dataset, output_folder=None):
         """
         Create evaluator(s) for a given dataset.
@@ -56,9 +46,8 @@ class Trainer(DefaultTrainer):
         dump_train = config.GLOBAL.DUMP_TRAIN
         return build_evaluator(cfg, dataset_name, dataset, output_folder, dump=dump_train)
 
-    @classmethod
     def test_with_TTA(cls, cfg, model):
-        logger = logging.getLogger("cvpods.trainer")
+        logger = logging.getLogger("cvpack2.trainer")
         # In the end of training, run an evaluation with TTA
         # Only support some R-CNN models.
         logger.info("Running inference with test-time augmentation ...")
@@ -67,29 +56,23 @@ class Trainer(DefaultTrainer):
         res = OrderedDict({k + "_TTA": v for k, v in res.items()})
         return res
 
+    cls.build_evaluator = build_evaluator
+    cls.test_with_TTA = test_with_TTA
 
-def stage_main(args, cfg, build):
-    cfg.merge_from_list(args.opts)
-    cfg, logger = default_setup(cfg, args)
-    model_build_func = build
+    return cls
+
+
+def main(args):
+    config.merge_from_list(args.opts)
+    cfg, logger = default_setup(config, args)
 
     """
     If you'd like to do anything fancier than the standard training logic,
     consider writing your own training loop or subclassing the trainer.
     """
-    trainer = Trainer(cfg, model_build_func)
-    trainer.resume_or_load(resume=args.resume)
+    trainer = trainer_decrator(RUNNERS.get(cfg.TRAINER.TYPE))(cfg, model_builder=build_model)
 
-    if args.eval_only:
-        DefaultCheckpointer(
-            trainer.model, save_dir=cfg.OUTPUT_DIR, resume=args.resume).resume_or_load(
-                cfg.MODEL.WEIGHTS, resume=args.resume)
-        res = Trainer.test(cfg, trainer.model)
-        if comm.is_main_process():
-            verify_results(cfg, res)
-        if cfg.TEST.AUG.ENABLED:
-            res.update(Trainer.test_with_TTA(cfg, trainer.model))
-        return res
+    trainer.resume_or_load(resume=args.resume)
 
     # check wheather worksapce has enough storeage space
     # assume that a single dumped model is 700Mb
@@ -105,61 +88,18 @@ def stage_main(args, cfg, build):
             [hooks.EvalHook(0, lambda: trainer.test_with_TTA(cfg, trainer.model))]
         )
 
+    logger.info("Running with full config:\n{}".format(cfg))
+    base_config = cfg.__class__.__base__()
+    logger.info("different config with base class:\n{}".format(cfg.diff(base_config)))
+
     trainer.train()
-
-    if comm.is_main_process() and cfg.MODEL.AS_PRETRAIN:
-        # convert last ckpt to pretrain format
-        convert_to_pretrained_model(
-            input=os.path.join(cfg.OUTPUT_DIR, "model_final.pth"),
-            save_path=os.path.join(cfg.OUTPUT_DIR, "model_final_pretrain_weight.pkl")
-        )
-
-
-def convert_to_pretrained_model(input, save_path):
-    obj = torch.load(input, map_location="cpu")
-    obj = obj["model"]
-
-    newmodel = {}
-    for k, v in obj.items():
-        if not k.startswith("encoder_q.") and not k.startswith("network"):
-            continue
-        old_k = k
-        if k.startswith("encoder_q."):
-            k = k.replace("encoder_q.", "")
-        elif k.startswith("network"):
-            k = k.replace("network.", "")
-        print(old_k, "->", k)
-        newmodel[k] = v.numpy()
-
-    res = {
-        "model": newmodel,
-        "__author__": "MOCO" if k.startswith("encoder_q.") else "CLS",
-        "matching_heuristics": True
-    }
-
-    with open(save_path, "wb") as f:
-        pkl.dump(res, f)
-
-
-def main(args):
-    if isinstance(config, list):
-        assert isinstance(build_model, list) and len(config) == len(build_model)
-        for cfg, build in zip(config, build_model):
-            stage_main(args, cfg, build)
-    else:
-        stage_main(args, config, build_model)
 
 
 if __name__ == "__main__":
     args = default_argument_parser().parse_args()
-    if isinstance(config, list):
-        assert len(config) > 0
-        print("soft link first config in list to {}".format(config[0].OUTPUT_DIR))
-        config[0].link_log()
-    else:
-        print("soft link to {}".format(config.OUTPUT_DIR))
-        config.link_log()
     print("Command Line Args:", args)
+    config.link_log()
+    print("soft link to {}".format(config.OUTPUT_DIR))
     launch(
         main,
         args.num_gpus,
